pyproject.toml
[project]
name = "opencodememory"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0",
    "gitpython>=3.1",
    "click>=8.0",
    "python-dateutil>=2.9",
]

[project.scripts]
ocm = "ocm.install.cli:main"
ocm-hook = "ocm.hooks.handler:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ocm"]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]
